Switched to BigInteger for values

+++ FibonacciPro.Console/FibonacciCalculator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -13,7 +14,7 @@ namespace FibonacciPro.ConsoleApplication
         /// </summary>
         /// <param name="n">number of digits of the fibonacci sequence to compute</param>
         /// <returns>an array containing the first n numbers in the fibonacci sequence</returns>
-        public static double[] Calculate(int n)
+        public static BigInteger[] Calculate(int n)
         {
             if (n <= 0)
                 throw new ArgumentException("n must be greater than 0", "n");
@@ -21,16 +22,16 @@ namespace FibonacciPro.ConsoleApplication
             //By definition
             if (n == 1)
             {
-                return new double[1] { 0 };
+                return new BigInteger[1] { 0 };
             }
 
             //By definition
             if (n == 2)
             {
-                return new double[2] { 0, 1 };
+                return new BigInteger[2] { 0, 1 };
             }
 
-            var result = new double[n];
+            var result = new BigInteger[n];
 
             var a = result[0] = 0;
             var b = result[1] = 1;
