+++ FibonacciPro.Console/IO/GenericIOHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,11 +9,11 @@ namespace FibonacciPro.ConsoleApplication.IO
 {
     public class GenericIOHandler : IInputHandler, IOutputHandler {
     
-        public Action<double[]> OutputHandler { get; set; }
+        public Action<BigInteger[]> OutputHandler { get; set; }
 
         public Func<int> InputHandler { get; set; } 
 
-        public void Write(double[] results)
+        public void Write(BigInteger[] results)
         {
             OutputHandler(results);
         }
