+++ FibonacciPro.Console/IO/IOutputHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,6 +9,6 @@ namespace FibonacciPro.ConsoleApplication.IO
 {
     public interface IOutputHandler
     {
-        void Write(double[] results);
+        void Write(BigInteger[] results);
     }
 }
